@@ -1,50 +1,45 @@
-import { Component } from "react";
+import { useState } from "react";
 import AppHeader from "../appHeader/AppHeader";
 import RandomChar from "../randomChar/RandomChar";
 import CharList from "../charList/CharList";
 import CharInfo from "../charInfo/CharInfo";
 import ErrorBoundery from "../errorBoundery/ErrorBoundery";
-
+import ComicsList from "../comicsList/ComicsList"
+ 
 import decoration from '../../resources/img/vision.png';
 
-class App extends Component {
-    state= {
-        showRandomChar: true,
-        selectedChar: null
-    }
+const  App = () => {
+    const [selectedChar, setChar] = useState(null);
+    const [showRandomChar, setRandomChar] = useState(true)
 
-    onCharSelected = (id) => {
-        this.setState({ 
-            selectedChar: id
-        })
+    const onCharSelected = (id) => {        
+            setChar(id)   
     }    
 
-    toggleRandomChar = () => {
-        this.setState((state) => {
-            return{
-                showRandomChar: !state.showRandomChar
-            }            
-        }) 
-    }
-
-    render(){
+    const toggleRandomChar = () => { 
+        setRandomChar(
+            !showRandomChar 
+        )   
+    }   
           
-        return (
-            <div className="app">
-                <AppHeader/>
-                <main>   
-                    {this.state.showRandomChar ? <RandomChar/> : null}
-                    <div className="char__content">
-                        <CharList onCharSelected={this.onCharSelected}/>
-                        <ErrorBoundery>
-                            <CharInfo charId={this.state.selectedChar}/>
-                        </ErrorBoundery>                        
-                    </div>
-                    <img className="bg-decoration" src={decoration} alt="vision"/>
-                </main>
-            </div>
-        )
-    }    
+    return (
+        <div className="app">
+            <AppHeader/>
+            <main>   
+                <ErrorBoundery>
+                    <RandomChar/>
+                </ErrorBoundery>
+                <div className="char__content">
+                    <CharList onCharSelected={onCharSelected}/>
+                    {/* <ComicsList/> */}
+                    <ErrorBoundery>
+                        <CharInfo charId={selectedChar}/>
+                    </ErrorBoundery>                        
+                </div>
+                <img className="bg-decoration" src={decoration} alt="vision"/>
+            </main>
+        </div>
+    )      
 }
 
 export default App;
