@@ -1,69 +1,73 @@
+import { useState, useEffect } from 'react';
+import useMarvelService from '../../services/MarvelService';
 import './comicsList.scss';
-import uw from '../../resources/img/UW.png';
-import xMen from '../../resources/img/x-men.png';
+import ErrorMessage from '../erorrMessage/errorMessage';
+import Spinner from "../spinner/spinner"
+
 
 const ComicsList = () => {
+    const [comicList, setComicList] = useState([])
+    const [newItemLoading, setNewItemLoading] = useState(false)
+    const [offset, setOffset] = useState(0)
+    const [comicsEnded, setComicsEnded] = useState(false);
+
+    const {loading, error, getAllComics} = useMarvelService();
+    
+    useEffect(() => {
+        onRequest(offset, true)
+    })
+
+    const onRequest = (offset, initial) => {
+        initial ? setNewItemLoading(false) : setNewItemLoading(true)
+        getAllComics(offset)
+            .then(onComicListLoaded)
+    }
+
+    const onComicListLoaded = (newComicsList) => {
+        let ended = false;
+        if(newComicsList.length < 8){
+            ended = true
+        }
+        setComicList([...comicList, ...newComicsList])
+        setNewItemLoading(false)
+        setOffset(offset + 8)
+        setComicsEnded(ended)
+    }
+
+    function renderItems (arr) {
+        const items = arr.map((item, i) => {
+            return (
+                <li className="comics__item" key={i}>
+                    <a href="#">
+                        <img src={item.thumbnail} alt={item.title} className="comics__item-img"/>
+                        <div className="comics__item-name">{item.title}</div>
+                        <div className="comics__item-price">{item.price}</div>
+                    </a>
+                </li>
+            )
+        })
+        return (
+            <ul className="comics__grid">
+                {items}
+            </ul>
+        )
+    }
+
+    const items = renderItems(comicList)
+
+    const errorMessage = error ? <ErrorMessage/> : null;
+    const spinner = loading && !newItemLoading ? <Spinner/> : null;
+    
     return (
         <div className="comics__list">
-            <ul className="comics__grid">
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={uw} alt="ultimate war" className="comics__item-img"/>
-                        <div className="comics__item-name">ULTIMATE X-MEN VOL. 5: ULTIMATE WAR TPB</div>
-                        <div className="comics__item-price">9.99$</div>
-                    </a>
-                </li>
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={xMen} alt="x-men" className="comics__item-img"/>
-                        <div className="comics__item-name">X-Men: Days of Future Past</div>
-                        <div className="comics__item-price">NOT AVAILABLE</div>
-                    </a>
-                </li>
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={uw} alt="ultimate war" className="comics__item-img"/>
-                        <div className="comics__item-name">ULTIMATE X-MEN VOL. 5: ULTIMATE WAR TPB</div>
-                        <div className="comics__item-price">9.99$</div>
-                    </a>
-                </li>
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={xMen} alt="x-men" className="comics__item-img"/>
-                        <div className="comics__item-name">X-Men: Days of Future Past</div>
-                        <div className="comics__item-price">NOT AVAILABLE</div>
-                    </a>
-                </li>
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={uw} alt="ultimate war" className="comics__item-img"/>
-                        <div className="comics__item-name">ULTIMATE X-MEN VOL. 5: ULTIMATE WAR TPB</div>
-                        <div className="comics__item-price">9.99$</div>
-                    </a>
-                </li>
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={xMen} alt="x-men" className="comics__item-img"/>
-                        <div className="comics__item-name">X-Men: Days of Future Past</div>
-                        <div className="comics__item-price">NOT AVAILABLE</div>
-                    </a>
-                </li>
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={uw} alt="ultimate war" className="comics__item-img"/>
-                        <div className="comics__item-name">ULTIMATE X-MEN VOL. 5: ULTIMATE WAR TPB</div>
-                        <div className="comics__item-price">9.99$</div>
-                    </a>
-                </li>
-                <li className="comics__item">
-                    <a href="#">
-                        <img src={xMen} alt="x-men" className="comics__item-img"/>
-                        <div className="comics__item-name">X-Men: Days of Future Past</div>
-                        <div className="comics__item-price">NOT AVAILABLE</div>
-                    </a>
-                </li>
-            </ul>
-            <button className="button button__main button__long">
+            {errorMessage}
+            {spinner}
+            {items}
+            <button
+                disabled={newItemLoading}
+                style={{'display' : comicsEnded ? 'none' : 'block'}}
+                onClick={() => onRequest(offset)}
+                className="button button__main button__long">
                 <div className="inner">load more</div>
             </button>
         </div>
