@@ -1,8 +1,8 @@
 import img from './error.gif'
-const ErorrMessage = () => {
+const ErrorMessage = () => {
    return (
       <img style={{ display: 'block', width: "250px", height: "250px",objectFit: 'contain', margin: "0 auto"}}
-      src={img} alt="Erorr"/>
+      src={img} alt="Error"/>
    )
 }
-export default ErorrMessage;
+export default ErrorMessage;
