@@ -1,87 +1,65 @@
-import { Component } from 'react';
+import { useState, useEffect } from 'react';
 import './randomChar.scss';
 import mjolnir from '../../resources/img/mjolnir.png';
-import MarvelService from '../../services/MarvelService';
+import useMarvelService from '../../services/MarvelService';
 import ErorrMessage from '../erorrMessage/errorMessage';
 import Spinner from '../spinner/spinner';
 
-class RandomChar extends Component {
-    state = {
-        char: {},
-        loading: true,
-        error: false
-    }
-    marvelService = new MarvelService()
-    componentDidMount(){        
-        this.updateChar()
-        // this.timerId = setInterval(this.updateChar, 4000)
-    }
-    componentWillUnmount(){
-        clearInterval(this.timerId)
+const  RandomChar = () => {
+    const [char, setChar] = useState(null);
+    const {loading, error, getCharacter, clearError} = useMarvelService()
+
+    useEffect(() => {
+        updateChar();
+        // const timerId = setInterval(updateChar, 60000);
+
+        return () => {
+            // clearInterval(timerId)
+        }
+    }, [])
+
+    const onCharLoaded = (char) => {        
+        setChar(char);
     }
 
-    onCharLoaded = (char) => {
-        this.setState(
-            {char, 
-                loading: false
-            }
-        )
-    }
-    onCharLoading = () => {
-        this.setState({
-            loading: true
-        })
-    } 
 
-    onErorr = () => {
-        this.setState({ 
-            loading: false, 
-            error: true
-         })
-    }
-    updateChar = () => {
-        const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000)
-        this.onCharLoading()
-        this.marvelService
-            .getCharacter(id)
-            .then(this.onCharLoaded)
-            .catch(this.onErorr)
-    }
-    render(){
-        const {char, loading, error} = this.state;
-        const erorrMessage = error ? <ErorrMessage/> : null;
-        const spinner = loading ? <Spinner/> : null;
-        const content = !(loading || error) ? <View char={char}/> : null
+    const updateChar = () => {
+        clearError()
+        const id = Math.floor(Math.random() * (1011400 - 1011000)) + 1011000;        
+        getCharacter(id)
+            .then(onCharLoaded)            
+    }   
+    
+    const erorrMessage = error ? <ErorrMessage/> : null;
+    const spinner = loading ? <Spinner/> : null;
+    const content = !(loading || error || !char) ? <View char={char}/> : null
 
-        return (
-            <div className="randomchar">
-                {erorrMessage}
-                {spinner}
-                {content}
-                <div className="randomchar__static">
-                    <p className="randomchar__title">
-                        Random character for today!<br/>
-                        Do you want to get to know him better?
-                    </p>
-                    <p className="randomchar__title">
-                        Or choose another one
-                    </p>
-                    <button className="button button__main"
-                        onClick={this.updateChar}>
-                        <div className="inner">try it</div>
-                    </button>
-                    <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
-                </div>
+    return (
+        <div className="randomchar">
+            {erorrMessage}
+            {spinner}
+            {content}
+            <div className="randomchar__static">
+                <p className="randomchar__title">
+                    Random character for today!<br/>
+                    Do you want to get to know him better?
+                </p>
+                <p className="randomchar__title">
+                    Or choose another one
+                </p>
+                <button className="button button__main"
+                    onClick={updateChar}>
+                    <div className="inner">try it</div>
+                </button>
+                <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
             </div>
-        )
-    }
+        </div>
+    )
     
 }
 
-const View = ({char}) => {
-    
+const View = ({char}) => {    
     const {name, description, thumbnail, homepage, wiki} = char
-
     let imgStyle = {'objectFit' : 'cover'};
     if (thumbnail === 'http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available.jpg') {
         imgStyle = {'objectFit' : 'contain'};
